requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.3
streamlit==1.53.0
plotly==5.24.1
